Conflicts:
	src/test/Test.cpp

+++ src/core/moves/SequenctialMoveSchedule.h
@@ -34,8 +34,8 @@ namespace RevBayesCore {
         // pure virtual public methods
         SequentialMoveSchedule*                         clone(void) const;
         double                                          getNumberMovesPerIteration(void) const;
-        Move*                                           nextMove(size_t g);
-
+        Move*                                           nextMove(unsigned long g);
+        
     private:
         
         // Hidden member variables
